some working ECIES crypto using spongycastle
caveat: spongycastle is only for Android, while bouncycastle is not for Android. Spongycastle is not "signed by a JCA Code Singning Certification Authority" cause on Android nobody would care. Sun-Java on the other hand does care. For years it was considered a good idea to just stick with Sun-Java to avoid weird issues but here in this case, openjdk also skipps the certificate test, so I recommend to use openJDK for now or find the switch to disable the signature check.